its getting there

+++ goal_achieving/DesireCategoryAsParameter.py
@@ -1,3 +1,5 @@
+# CREATED JUNE OF 2023 BY SPENCER CHAVEZ
+
 from pydantic import BaseModel
 from pydantic.fields import Optional
 
@@ -5,6 +7,7 @@ from pydantic.fields import Optional
 class DesireCategoryAsParameter(BaseModel):
     name: Optional[str]
     userId: Optional[int]
-    colorR: Optional[int]
-    colorB: Optional[int]
-    colorG: Optional[int]
+    # colors are in the range of 0.0 to 1.0
+    colorR: Optional[float]
+    colorB: Optional[float]
+    colorG: Optional[float]
